+++ Kite-SDK/PSPrintSDK/CI-ViewController.m
@@ -375,11 +375,7 @@ static NSString *const kApplePayBusinessName = @"Kite.ly"; //Replace with your b
     
     [self addCatsAndDogsImagePickersToKite:vc];
     
-    KITAssetsPickerController *customVc = [[KITAssetsPickerController alloc] init];
-    self.customDataSources = @[[[CustomAssetCollectionDataSource alloc] init]];
-    customVc.collectionDataSources = self.customDataSources;
-    
-    [vc addCustomPhotoProviderWithViewController:(UIViewController<OLCustomPickerController> *)customVc name:@"External" icon:[UIImage imageNamed:@"cat"] prepopulatedAssets:assets];
+    [vc addCustomPhotoProviderWithViewController:nil name:@"External" icon:[UIImage imageNamed:@"cat"] prepopulatedAssets:assets];
     
     [self presentViewController:vc animated:YES completion:NULL];
 }
@@ -394,15 +390,19 @@ static NSString *const kApplePayBusinessName = @"Kite.ly"; //Replace with your b
     vc.disableCameraRoll = YES;
     [OLKitePrintSDK setInstagramEnabledWithClientID:@"" secret:@"" redirectURI:@""];
     
-    KITAssetsPickerController *customVc = [[KITAssetsPickerController alloc] init];
-    self.customDataSources = @[[[CustomAssetCollectionDataSource alloc] init]];
-    customVc.collectionDataSources = self.customDataSources;
-    
-    [vc addCustomPhotoProviderWithViewController:(UIViewController<OLCustomPickerController> *)customVc name:@"External" icon:[UIImage imageNamed:@"cat"] prepopulatedAssets:assets];
+    [vc addCustomPhotoProviderWithViewController:nil name:@"External" icon:[UIImage imageNamed:@"cat"] prepopulatedAssets:assets];
     
     [ipvc dismissViewControllerAnimated:YES completion:^{
         [self presentViewController:vc animated:YES completion:NULL];
     }];
+}
+
+- (UIViewController<OLCustomPickerController> *_Nonnull)imagePickerViewControllerForName:(NSString *_Nonnull)name{
+    KITAssetsPickerController *customVc = [[KITAssetsPickerController alloc] init];
+    self.customDataSources = @[[[CustomAssetCollectionDataSource alloc] init]];
+    customVc.collectionDataSources = self.customDataSources;
+    
+    return (UIViewController<OLCustomPickerController> *)customVc;
 }
 
 @end