+++ Kite-SDK/PSPrintSDK/OLFrameOrderReviewViewController.m
@@ -98,6 +98,9 @@ CGFloat innerMargin = 3;
         self.vcDelegateForCustomVc = vc; //Keep strong reference
         vc.providerForPresentedVc = [OLUserSession currentSession].kiteVc.customImageProviders.firstObject;
         UIViewController<OLCustomPickerController> *customVc = [(OLCustomViewControllerPhotoProvider *)[OLUserSession currentSession].kiteVc.customImageProviders.firstObject vc];
+        if (!customVc){
+            customVc = [[OLUserSession currentSession].kiteVc.delegate imagePickerViewControllerForName:vc.providerForPresentedVc.name];
+        }
         [customVc safePerformSelector:@selector(setDelegate:) withObject:vc];
         [customVc safePerformSelector:@selector(setProductId:) withObject:self.product.templateId];
         [customVc safePerformSelector:@selector(setSelectedAssets:) withObject:[[NSMutableArray alloc] init]];