+++ Kite-SDK/PSPrintSDK/OLPaymentViewController.m
@@ -972,14 +972,6 @@ UIActionSheetDelegate, UITextFieldDelegate, UINavigationControllerDelegate, UITa
     }];
 }
 
-- (BOOL)showPhoneEntryField {
-    if ([self.kiteDelegate respondsToSelector:@selector(shouldShowPhoneEntryOnCheckoutScreen)]) {
-        return (![OLUserSession currentSession].kiteVc.hidePhoneEntryOnCheckoutScreen);
-    }
-    
-    return [OLKiteABTesting sharedInstance].requirePhoneNumber;
-}
-
 #pragma mark Button Actions
 
 - (IBAction)onButtonApplyPromoCodeClicked:(id)sender {
@@ -1110,11 +1102,7 @@ UIActionSheetDelegate, UITextFieldDelegate, UINavigationControllerDelegate, UITa
         
         [lineItems addObject:[PKPaymentSummaryItem summaryItemWithLabel:[OLKitePrintSDK applePayPayToString] amount:[cost totalCostInCurrency:self.printOrder.currencyCode]]];
         paymentRequest.paymentSummaryItems = lineItems;
-        NSUInteger requiredFields = PKAddressFieldPostalAddress | PKAddressFieldName | PKAddressFieldEmail;
-        if ([self showPhoneEntryField]){
-            requiredFields = requiredFields | PKAddressFieldPhone;
-        }
-        paymentRequest.requiredShippingAddressFields = requiredFields;
+        paymentRequest.requiredShippingAddressFields = PKAddressFieldPostalAddress | PKAddressFieldName | PKAddressFieldEmail | PKAddressFieldPhone;
         UIViewController *paymentController;
         paymentController = [[PKPaymentAuthorizationViewController alloc]
                              initWithPaymentRequest:paymentRequest];
@@ -1242,7 +1230,7 @@ UIActionSheetDelegate, UITextFieldDelegate, UINavigationControllerDelegate, UITa
     if (self.printOrder.jobs.count == 0){
         return;
     }
-    if (![self checkForShippingAddress]){
+    if ((selectedPaymentMethod == kOLPaymentMethodPayPal || selectedPaymentMethod == kOLPaymentMethodCreditCard) && ![self checkForShippingAddress]){
         return;
     }